Split crypto.rs into smaller private modules
