+++ rs-matter/src/crypto.rs
@@ -17,483 +17,18 @@
 
 //! Cryptographic abstractions and backend.
 
-use core::fmt::Debug;
-
 use embassy_sync::blocking_mutex::raw::RawMutex;
+
+use crate::error::Error;
 
 pub use rand_core::{CryptoRng, CryptoRngCore, RngCore};
 
-use crate::error::{Error, ErrorCode};
-use crate::tlv::{FromTLV, TLVElement, TLVTag, TLVWrite, ToTLV, TLV};
-use crate::utils::cell::RefCell;
-use crate::utils::init::{init, zeroed, Init, IntoFallibleInit};
-use crate::utils::sync::blocking::Mutex;
-
-pub mod dummy;
-#[cfg(feature = "mbedtls")]
-pub mod mbedtls;
-#[cfg(feature = "openssl")]
-pub mod openssl;
-#[cfg(feature = "rustcrypto")]
-pub mod rustcrypto;
-
-/// Length of the hash returned by the hasher (`Crypto::hash`) in bytes.
-///
-/// As per the Matter spec, the hasher should be SHA-256.
-pub const HASH_LEN: usize = 32;
-
-/// Length of the HMAC hash returned by the HMAC hasher (`Crypto::hmac`) in bytes.
-///
-/// As per the Matter spec, the HMAC hasher should be HMAC-SHA-256.
-pub const HMAC_HASH_LEN: usize = HASH_LEN;
-
-/// Length of the canonical representation of an EC scalar in bytes.
-///
-/// As per the Matter spec, the curve used is secp256r1 (NIST P-256).
-pub const EC_CANON_SCALAR_LEN: usize = 32;
-
-/// Length of the canonical representation of an EC point in bytes.
-///
-/// As per the Matter spec, the curve used is secp256r1 (NIST P-256).
-pub const EC_CANON_POINT_LEN: usize = EC_CANON_SCALAR_LEN * 2 + 1;
-
-/// Length of the canonical representation of a public key in bytes.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// Note that this is the same as `EC_CANON_POINT_LEN`.
-pub const PKC_CANON_PUBLIC_KEY_LEN: usize = EC_CANON_POINT_LEN;
-
-/// Length of the canonical representation of a secret key in bytes.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// Note that this is the same as `EC_CANON_SCALAR_LEN`.
-pub const PKC_CANON_SECRET_KEY_LEN: usize = EC_CANON_SCALAR_LEN;
-
-/// Length of the canonical representation of a signature in bytes.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// Note that this is `2 * EC_CANON_SCALAR_LEN`, as the signature contains
-/// the (r, s) scalars computed using ECDSA.
-pub const PKC_SIGNATURE_LEN: usize = PKC_CANON_SECRET_KEY_LEN * 2;
-
-/// Length of the canonical representation of a shared secret in bytes.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// The shared secret is the ECDH computed value.
-pub const PKC_SHARED_SECRET_LEN: usize = 32;
-
-/// Length of the canonical representation of a 320-bit unsigned integer in bytes.
-///
-/// As per the Matter spec, this is used in the SPAKE2+ protocol.
-pub const UINT320_CANON_LEN: usize = 40;
-
-/// Length of the canonical representation of an AEAD key in bytes.
-///
-/// As per the Matter spec, the AEAD algorithm used is AES-CCM with 128-bit keys.
-pub const AEAD_CANON_KEY_LEN: usize = 16;
-
-/// Length of the nonce used in AEAD operations in bytes.
-///
-/// As per the Matter spec, the AEAD algorithm used is AES-CCM with a 13-byte nonce.
-pub const AEAD_NONCE_LEN: usize = 13;
-
-/// Length of the tag produced by AEAD operations in bytes.
-///
-/// As per the Matter spec, the AEAD algorithm used is AES-CCM with a 16-byte tag.
-pub const AEAD_TAG_LEN: usize = 16;
-
-/// Canonical representation of a hash value.
-///
-/// As per the Matter spec, the hasher should be SHA-256.
-pub type Hash = CryptoSensitive<HASH_LEN>;
-
-pub type HashRef<'a> = CryptoSensitiveRef<'a, HASH_LEN>;
-
-/// Zeroed hash value.
-pub const HASH_ZEROED: Hash = Hash::new();
-
-/// Canonical representation of an HMAC hash value.
-///
-/// As per the Matter spec, the HMAC hasher should be HMAC-SHA-256.
-pub type HmacHash = CryptoSensitive<HMAC_HASH_LEN>;
-
-pub type HmacHashRef<'a> = CryptoSensitiveRef<'a, HMAC_HASH_LEN>;
-
-/// Zeroed HMAC hash value.
-pub const HMAC_HASH_ZEROED: Hash = HmacHash::new();
-
-/// Canonical representation of a 320-bit unsigned integer (BE format).
-pub type CanonUint320 = CryptoSensitive<UINT320_CANON_LEN>;
-
-pub type CanonUint320Ref<'a> = CryptoSensitiveRef<'a, UINT320_CANON_LEN>;
-
-/// Zeroed 320-bit unsigned integer.
-pub const UINT320_ZEROED: CanonUint320 = CanonUint320::new();
-
-/// Canonical representation of an AEAD key.
-///
-/// As per the Matter spec, the AEAD algorithm used is AES-CCM with 128-bit keys.
-pub type CanonAeadKey = CryptoSensitive<AEAD_CANON_KEY_LEN>;
-
-pub type CanonAeadKeyRef<'a> = CryptoSensitiveRef<'a, AEAD_CANON_KEY_LEN>;
-
-/// Zeroed AEAD key.
-pub const AEAD_KEY_ZEROED: CanonAeadKey = CanonAeadKey::new();
-
-/// Canonical representation of an AEAD nonce.
-///
-/// As per the Matter spec, the AEAD algorithm used is AES-CCM with a 13-byte nonce.
-pub type AeadNonce = CryptoSensitive<AEAD_NONCE_LEN>;
-
-pub type AeadNonceRef<'a> = CryptoSensitiveRef<'a, AEAD_NONCE_LEN>;
-
-/// Zeroed AEAD nonce.
-pub const AEAD_NONCE_ZEROED: AeadNonce = AeadNonce::new();
-
-/// Canonical representation of an AEAD tag.
-///
-/// As per the Matter spec, the AEAD algorithm used is AES-CCM with a 16-byte tag.
-pub type AeadTag = CryptoSensitive<AEAD_TAG_LEN>;
-
-pub type AeadTagRef<'a> = CryptoSensitiveRef<'a, AEAD_TAG_LEN>;
-
-/// Zeroed AEAD tag.
-pub const AEAD_TAG_ZEROED: AeadTag = AeadTag::new();
-
-/// Canonical representation of an EC scalar.
-///
-/// As per the Matter spec, the curve used is secp256r1 (NIST P-256).
-pub type CanonEcScalar = CryptoSensitive<EC_CANON_SCALAR_LEN>;
-
-pub type CanonEcScalarRef<'a> = CryptoSensitiveRef<'a, EC_CANON_SCALAR_LEN>;
-
-/// Zeroed EC scalar.
-pub const EC_SCALAR_ZEROED: CanonEcScalar = CanonEcScalar::new();
-
-/// Canonical representation of an EC point (one byte prefix and then two affine scalar coordinates),
-/// uncompressed.
-///
-/// As per the Matter spec, the curve used is secp256r1 (NIST P-256).
-pub type CanonEcPoint = CryptoSensitive<EC_CANON_POINT_LEN>;
-
-pub type CanonEcPointRef<'a> = CryptoSensitiveRef<'a, EC_CANON_POINT_LEN>;
-
-/// Zeroed EC point.
-pub const EC_POINT_ZEROED: CanonEcPoint = CanonEcPoint::new();
-
-/// Canonical representation of a Public Key Cryptography secret key.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// Note that this is the same as `CanonEcScalar`.
-pub type CanonPkcSecretKey = CanonEcScalar;
-
-pub type CanonPkcSecretKeyRef<'a> = CanonEcScalarRef<'a>;
-
-/// Zeroed PKC secret key.
-pub const PKC_SECRET_KEY_ZEROED: CanonPkcSecretKey = EC_SCALAR_ZEROED;
-
-/// Canonical representation of a Public Key Cryptography public key.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// Note that this is the same as `CanonEcPoint`.
-pub type CanonPkcPublicKey = CanonEcPoint;
-
-pub type CanonPkcPublicKeyRef<'a> = CanonEcPointRef<'a>;
-
-/// Zeroed PKC public key.
-pub const PKC_PUBLIC_KEY_ZEROED: CanonPkcPublicKey = EC_POINT_ZEROED;
-
-/// Canonical representation of a Public Key Cryptography signature.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// Note that this is `2 * EC_CANON_SCALAR_LEN`, as the signature contains
-/// the (r, s) scalars computed using ECDSA.
-pub type CanonPkcSignature = CryptoSensitive<PKC_SIGNATURE_LEN>;
-
-pub type CanonPkcSignatureRef<'a> = CryptoSensitiveRef<'a, PKC_SIGNATURE_LEN>;
-
-/// Zeroed PKC signature.
-pub const EC_SIGNATURE_ZEROED: CanonPkcSignature = CryptoSensitive::new();
-
-/// Canonical representation of a Public Key Cryptography shared secret.
-///
-/// As per the Matter spec, the used Public Key Cryptograqphy should be
-/// Elliptic-Curve based, and specifically secp256r1 (NIST P-256).
-///
-/// The shared secret is the ECDH computed value.
-pub type CanonPkcSharedSecret = CryptoSensitive<PKC_SHARED_SECRET_LEN>;
-
-pub type CanonPkcSharedSecretRef<'a> = CryptoSensitiveRef<'a, PKC_SHARED_SECRET_LEN>;
-
-/// Zeroed PKC shared secret.
-pub const PKC_SHARED_SECRET_ZEROED: CanonPkcSharedSecret = CryptoSensitive::new();
-
-#[derive(Clone)]
-pub struct CryptoSensitive<const N: usize> {
-    data: [u8; N],
-}
-
-impl<const N: usize> CryptoSensitive<N> {
-    #[inline(always)]
-    pub const fn new() -> Self {
-        Self { data: [0u8; N] }
-    }
-
-    pub const fn new_from_ref(other: CryptoSensitiveRef<'_, N>) -> Self {
-        let mut this = Self::new();
-
-        this.load(other);
-
-        this
-    }
-
-    #[inline(always)]
-    pub fn init() -> impl Init<Self> {
-        init!(Self {
-            data <- zeroed(),
-        })
-    }
-
-    pub fn zeroize(&mut self) {
-        for b in &mut self.data {
-            *b = 0;
-        }
-    }
-
-    pub const fn reference(&self) -> CryptoSensitiveRef<'_, N> {
-        CryptoSensitiveRef::new(&self.data)
-    }
-
-    pub const fn load(&mut self, other: CryptoSensitiveRef<'_, N>) {
-        self.load_from_array(other.access());
-    }
-
-    pub const fn load_from_array(&mut self, data: &[u8; N]) {
-        self.data.copy_from_slice(data);
-    }
-
-    pub fn try_load_from_slice(&mut self, data: &[u8]) -> Result<(), Error> {
-        if data.len() != N {
-            return Err(ErrorCode::InvalidData.into());
-        }
-
-        self.data.copy_from_slice(data);
-
-        Ok(())
-    }
-
-    pub const fn access(&self) -> &[u8; N] {
-        &self.data
-    }
-
-    pub const fn access_mut(&mut self) -> &mut [u8; N] {
-        &mut self.data
-    }
-}
-
-impl<const N: usize> Drop for CryptoSensitive<N> {
-    fn drop(&mut self) {
-        self.zeroize();
-    }
-}
-
-impl<const N: usize> Default for CryptoSensitive<N> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<const N: usize> Debug for CryptoSensitive<N> {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "CryptoSensitive<{}>(**hidden**)", N)
-    }
-}
-
-#[cfg(feature = "defmt")]
-impl<const N: usize> defmt::Format for CryptoSensitive<N> {
-    fn format(&self, f: defmt::Formatter) {
-        defmt::write!(f, "CryptoSensitive<{}>(**hidden**)", N);
-    }
-}
-
-impl<const N: usize> From<CryptoSensitiveRef<'_, N>> for CryptoSensitive<N> {
-    fn from(other: CryptoSensitiveRef<'_, N>) -> Self {
-        let mut material = CryptoSensitive::new();
-
-        material.load(other);
-
-        material
-    }
-}
-
-impl<const N: usize> From<&[u8; N]> for CryptoSensitive<N> {
-    fn from(data: &[u8; N]) -> Self {
-        let mut material = CryptoSensitive::new();
-
-        material.load_from_array(data);
-
-        material
-    }
-}
-
-impl<const N: usize> From<[u8; N]> for CryptoSensitive<N> {
-    fn from(data: [u8; N]) -> Self {
-        let mut material = CryptoSensitive::new();
-
-        material.load_from_array(&data);
-
-        material
-    }
-}
-
-impl<const N: usize> TryFrom<&[u8]> for CryptoSensitive<N> {
-    type Error = Error;
-
-    fn try_from(data: &[u8]) -> Result<Self, Self::Error> {
-        let mut material = CryptoSensitive::new();
-
-        material.try_load_from_slice(data)?;
-
-        Ok(material)
-    }
-}
-
-impl<'a, const N: usize> FromTLV<'a> for CryptoSensitive<N> {
-    fn from_tlv(element: &TLVElement<'a>) -> Result<Self, crate::error::Error> {
-        Ok(Self {
-            data: element
-                .str()?
-                .try_into()
-                .map_err(|_| ErrorCode::ConstraintError)?,
-        })
-    }
-
-    fn init_from_tlv(element: TLVElement<'a>) -> impl Init<Self, Error> {
-        Init::chain(Self::init().into_fallible(), move |this| {
-            let data = element.str()?;
-            if data.len() != N {
-                Err(ErrorCode::ConstraintError)?;
-            }
-
-            this.access_mut().copy_from_slice(data);
-
-            Ok(())
-        })
-    }
-}
-
-impl<const N: usize> ToTLV for CryptoSensitive<N> {
-    fn to_tlv<W: TLVWrite>(&self, tag: &TLVTag, mut tw: W) -> Result<(), Error> {
-        tw.str(tag, &self.data)
-    }
-
-    fn tlv_iter(&self, tag: TLVTag) -> impl Iterator<Item = Result<TLV<'_>, Error>> {
-        TLV::str(tag, self.data.as_slice()).into_tlv_iter()
-    }
-}
-
-#[derive(Copy, Clone)]
-pub struct CryptoSensitiveRef<'a, const N: usize> {
-    data: &'a [u8; N],
-}
-
-impl<'a, const N: usize> CryptoSensitiveRef<'a, N> {
-    #[inline(always)]
-    pub const fn new(material: &'a [u8; N]) -> Self {
-        Self { data: material }
-    }
-
-    #[inline(always)]
-    pub fn new_from_slice(material: &'a [u8]) -> Self {
-        assert_eq!(material.len(), N);
-
-        Self::new(Self::as_array(material).unwrap()) // TODO
-    }
-
-    #[inline(always)]
-    pub fn try_new(material: &'a [u8]) -> Result<Self, Error> {
-        if material.len() != N {
-            Err(ErrorCode::InvalidData)?;
-        }
-
-        Ok(Self::new(Self::as_array(material).unwrap())) // TODO
-    }
-
-    pub fn split<const M1: usize, const M2: usize>(
-        &self,
-    ) -> (CryptoSensitiveRef<'a, M1>, CryptoSensitiveRef<'a, M2>) {
-        let (left, right) = self.data.split_at(M1);
-
-        (
-            CryptoSensitiveRef::new_from_slice(left),
-            CryptoSensitiveRef::new_from_slice(right),
-        )
-    }
-
-    pub const fn access(&self) -> &'a [u8; N] {
-        self.data
-    }
-
-    // TODO: `as_array` is not yet const fn in Rust core
-    const fn as_array<const L: usize>(slice: &'a [u8]) -> Option<&'a [u8; L]> {
-        if slice.len() == L {
-            let ptr = slice.as_ptr() as *const [u8; L];
-
-            // SAFETY: The underlying array of a slice can be reinterpreted as an actual array `[T; N]` if `N` is not greater than the slice's length.
-            let me = unsafe { &*ptr };
-            Some(me)
-        } else {
-            None
-        }
-    }
-}
-
-impl<const N: usize> Debug for CryptoSensitiveRef<'_, N> {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "CryptoSensitiveRef<{}>(**hidden**)", N)
-    }
-}
-
-#[cfg(feature = "defmt")]
-impl<const N: usize> defmt::Format for CryptoSensitiveRef<'_, N> {
-    fn format(&self, f: defmt::Formatter) {
-        defmt::write!(f, "CryptoSensitiveRef<{}>(**hidden**)", N);
-    }
-}
-
-impl<'a, const N: usize> From<&'a CryptoSensitive<N>> for CryptoSensitiveRef<'a, N> {
-    fn from(cs: &'a CryptoSensitive<N>) -> Self {
-        cs.reference()
-    }
-}
-
-impl<'a, const N: usize> From<&'a [u8; N]> for CryptoSensitiveRef<'a, N> {
-    fn from(data: &'a [u8; N]) -> Self {
-        Self::new(data)
-    }
-}
-
-impl<'a, const N: usize> TryFrom<&'a [u8]> for CryptoSensitiveRef<'a, N> {
-    type Error = Error;
-
-    fn try_from(data: &'a [u8]) -> Result<Self, Self::Error> {
-        Self::try_new(data)
-    }
-}
+pub use canon::*;
+pub use rand::*;
+
+pub mod backend;
+mod canon;
+mod rand;
 
 /// Trait representing a cryptographic backend.
 ///
@@ -1092,88 +627,6 @@ pub trait EcPoint<'a, const LEN: usize, const SCALAR_LEN: usize> {
     fn write_canon(&self, point: &mut CryptoSensitive<LEN>);
 }
 
-pub struct SharedRand<M: RawMutex, T> {
-    shared: Mutex<M, RefCell<T>>,
-}
-
-impl<M: RawMutex, T> SharedRand<M, T> {
-    pub const fn new(rand: T) -> Self {
-        Self {
-            shared: Mutex::new(RefCell::new(rand)),
-        }
-    }
-
-    pub fn init(rand: impl Init<T>) -> impl Init<Self> {
-        init!(Self {
-            shared <- Mutex::init(RefCell::init(rand)),
-        })
-    }
-}
-
-impl<M: RawMutex, T> rand_core::RngCore for &SharedRand<M, T>
-where
-    T: rand_core::RngCore,
-{
-    fn next_u32(&mut self) -> u32 {
-        self.shared.lock(|rand| rand.borrow_mut().next_u32())
-    }
-
-    fn next_u64(&mut self) -> u64 {
-        self.shared.lock(|rand| rand.borrow_mut().next_u64())
-    }
-
-    fn fill_bytes(&mut self, dest: &mut [u8]) {
-        self.shared.lock(|rand| rand.borrow_mut().fill_bytes(dest))
-    }
-
-    fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), rand_core::Error> {
-        self.shared
-            .lock(|rand| rand.borrow_mut().try_fill_bytes(dest))
-    }
-}
-
-impl<M: RawMutex, T> CryptoRng for &SharedRand<M, T> where T: CryptoRng {}
-
-pub struct WeakTestOnlyRand(u32);
-
-impl WeakTestOnlyRand {
-    const SEED: u32 = 2463534242;
-
-    pub const fn new_default() -> Self {
-        Self(Self::SEED)
-    }
-
-    pub const fn new(seed: u32) -> Self {
-        Self(seed)
-    }
-}
-
-impl RngCore for WeakTestOnlyRand {
-    fn next_u32(&mut self) -> u32 {
-        self.0 = self.0 ^ (self.0 << 13);
-        self.0 = self.0 ^ (self.0 >> 17);
-        self.0 = self.0 ^ (self.0 << 5);
-
-        self.0
-    }
-
-    fn next_u64(&mut self) -> u64 {
-        rand_core::impls::next_u64_via_u32(self)
-    }
-
-    fn fill_bytes(&mut self, dest: &mut [u8]) {
-        rand_core::impls::fill_bytes_via_next(self, dest)
-    }
-
-    fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), rand_core::Error> {
-        rand_core::impls::fill_bytes_via_next(self, dest);
-
-        Ok(())
-    }
-}
-
-impl CryptoRng for WeakTestOnlyRand {}
-
 #[allow(unused)]
 pub fn default_crypto<'s, M, R>(
     rand: R,
@@ -1184,19 +637,19 @@ where
     R: CryptoRngCore + 's,
 {
     #[cfg(feature = "openssl")]
-    let crypto = openssl::OpenSslCrypto::new(singleton_secret_key);
+    let crypto = backend::openssl::OpenSslCrypto::new(singleton_secret_key);
 
     #[cfg(all(feature = "mbedtls", not(feature = "openssl")))]
-    let crypto = mbedtls::MbedtlsCrypto::<M, _>::new(rand, singleton_secret_key);
+    let crypto = backend::mbedtls::MbedtlsCrypto::<M, _>::new(rand, singleton_secret_key);
 
     #[cfg(all(
         feature = "rustcrypto",
         not(any(feature = "openssl", feature = "mbedtls"))
     ))]
-    let crypto = rustcrypto::RustCrypto::<M, _>::new(rand, singleton_secret_key);
+    let crypto = backend::rustcrypto::RustCrypto::<M, _>::new(rand, singleton_secret_key);
 
     #[cfg(not(any(feature = "openssl", feature = "mbedtls", feature = "rustcrypto")))]
-    let crypto = dummy::DummyCrypto;
+    let crypto = backend::dummy::DummyCrypto;
 
     crypto
 }
